Updated the messages to include long long types for the time since they are not an int

+++ include/falcon-comms/autotuner_comms.hpp
@@ -27,7 +27,7 @@ public:
    * @param time current time
    * @return StateResponse if successful
    */
-  StateResponse subscribe_state_response(int timeout_ms, int time);
+  StateResponse subscribe_state_response(int timeout_ms, long long time);
 
 protected:
   NatsManager &hub_;
